feat: remove migrate command and migrate all on startup

+++ zip-plugin/src/main/java/net/imprex/zip/BackpackMigrator.java
@@ -1,17 +1,20 @@
 package net.imprex.zip;
 
+import java.awt.geom.IllegalPathStateException;
 import java.io.FileInputStream;
 import java.io.FileOutputStream;
+import java.io.IOException;
 import java.io.OutputStreamWriter;
 import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
 import java.nio.file.Path;
+import java.util.stream.Stream;
 
 import com.google.common.io.ByteStreams;
 import com.google.gson.JsonObject;
 
 import io.netty.buffer.Unpooled;
-import net.imprex.zip.common.BPKey;
+import net.imprex.zip.common.BPConstants;
 import net.imprex.zip.common.Ingrim4Buffer;
 import net.imprex.zip.common.UniqueId;
 import net.imprex.zip.common.ZIPLogger;
@@ -31,6 +34,50 @@ public class BackpackMigrator {
 	 * buffer.writeString(this.typeRaw);
 	 * buffer.writeByteArray(NmsInstance.itemstackToBinary(this.content));
 	 */
+	
+	public static void checkForMigrations(Path folderPath) {
+		long startTime = System.currentTimeMillis();
+		int statisticSuccessful = 0;
+		int statisticFailed = 0;
+	
+		ZIPLogger.info("Checking for migration data...");
+
+		try (Stream<Path> stream = Files.walk(folderPath, 1)) {
+			Path[] paths = stream
+					.filter(file -> !Files.isDirectory(file))
+					.filter(file -> Files.isRegularFile(file))
+					.filter(file -> !file.getFileName().toString().endsWith(".json"))
+					.toArray(Path[]::new);
+	
+			ZIPLogger.info("Migration found " + paths.length + " outdated backpacks.");
+			
+			if (paths.length == 0) {
+				return;
+			}
+			
+			for (Path file : paths) {
+				try {
+					UniqueId id = UniqueId.fromString(file.getFileName().toString());
+					if (BackpackMigrator.migrate(folderPath, id)) {
+						statisticSuccessful++;
+					} else {
+						statisticFailed++;
+					}
+				} catch (Exception e) {
+					e.printStackTrace();
+					statisticFailed++;
+				}
+			}
+			
+			ZIPLogger.info(String.format("Migration finished. %d/%d backpacks migrated and %d failed to migrate in %d seconds.",
+					paths.length,
+					statisticSuccessful,
+					statisticFailed,
+					Math.round((System.currentTimeMillis() - startTime) / 1000)));
+		} catch (IOException e) {
+			ZIPLogger.error("Error when migrating backpacks", e);
+		}
+	}
 
 	public static boolean migrate(Path folderPath, UniqueId id) {
 		try {
@@ -51,12 +98,16 @@ public class BackpackMigrator {
 			byte[] previousContent = buffer.readByteArray();
 			
 			JsonObject json = new JsonObject();
-			json.addProperty(BPKey.VERSION, 2);
-			json.addProperty(BPKey.ID, UniqueId.fromByteArray(previousId).toString());
-			json.addProperty(BPKey.TYPE_RAW, previousRawType);
-			json.add(BPKey.INVENTORY, NmsInstance.migrateToJsonElement(previousContent));
+			json.addProperty(BPConstants.KEY_VERSION, BPConstants.VERSION);
+			json.addProperty(BPConstants.KEY_ID, UniqueId.fromByteArray(previousId).toString());
+			json.addProperty(BPConstants.KEY_TYPE_RAW, previousRawType);
+			json.add(BPConstants.KEY_INVENTORY, NmsInstance.migrateToJsonElement(previousContent));
 
 			Path newFile = folderPath.resolve(id.toString() + ".json");
+			if (Files.exists(newFile)) {
+				throw new IllegalPathStateException("File path for migration " + id.toString() + " already exist!");
+			}
+			
 			try (FileOutputStream outputStream = new FileOutputStream(newFile.toFile());
 					OutputStreamWriter outputStreamWriter = new OutputStreamWriter(outputStream, StandardCharsets.UTF_8)) {
 				Backpack.GSON.toJson(json, outputStreamWriter);
